+++ apps/frontend/src/constants/infrastructure.ts
@@ -29,4 +29,14 @@ export const AMM_SERVICE = {
   [Environments.Testnet]: 'https://amm-apy.test.sovryn.app/',
 };
 
+export const INDEXER_SERVICE = {
+  [Environments.Mainnet]: 'https://indexer.sovryn.app/',
+  [Environments.Testnet]: 'https://indexer.test.sovryn.app/',
+};
+
 export const ESTIMATOR_URI = process.env.REACT_APP_ESTIMATOR_URI;
+
+export const SOVRYN_INDEXER = {
+  [Environments.Mainnet]: 'https://indexer.sovryn.app',
+  [Environments.Testnet]: 'https://indexer.test.sovryn.app',
+};